Move navigation component into a directory
